refactor: Puppeteer 삭제 후 코드 정리
