--- src/commands/end.js
@@ -1,5 +1,0 @@
-import { closeBrowser } from "../puppeteer/browser.js";
-
-export const stopDebugging = async () => {
-  closeBrowser();
-};
